Add files via upload
Hello everyone, i make a external tool for edit to heist cut%. It has not need to Cheat Engine support, just need your PC installed C++ Runtime library.

It can edit apartment heist(Only host valid),doomdays heist, casino heist's and cayo-perico heist's player cut%.

In addition to these, it can edit the Potential Take of the casino and the real take of the in mission, as well as the real take of Cayo-Perico.

Common Problem Q&A:

Q:When the casino heist mission is started, it crashes when selecting entrances and exits.
A:You need unlock all access points and POI before start the last mission. 

Q:When I started the casino mission, it got stuck on the splash screen. (Started the casino mission as the host for the first time and killed the NPC)
A：When you first time do casino heist, do not kill any NPC.

Q:Sell vehicle does not work, it does not work some version?
A:No,it work in all version(At least in the game 1.59 version, tools 3.2 work). Sell vehicle requires you to enter the selling option after entering the LSC, and modify it after the vehicle price appears in the in-game menu. After the modification.

Q:Sell vehicle shows invalid transaction.
A:You only can sell your personal vehicle.

Q：When setting up a vehicle for sale, it reminds me that I cannot exceed 8M?
A：This is my suggestion to you for safety reasons. In theory, you can continue to execute the value you set after confirming it (this may be a mistake when I wrote the code). But for security reasons, I still recommend you to set a value less than 8M again.
Q： Heist Job max payout?
A:
Casino all cost<3.66
cayo all cost<2.55
all cost=main cost+other cost.



Current known issues(For 3.1)(Resolve in 3.1.2):
1. Setup cut for apartment will crash(Tool's crash).
2.Some MessageBox text lost
3.Casino job not fully unlocked
4.Have some error label text...(Cayo player cut player ID text error)